refactor: 💡 move function in session storage
